Move ES client base classes to Source/common/es/
Move SNTEndpointSecurityClientBase, SNTEndpointSecurityClient,
SNTEndpointSecurityEventHandler, and SNTEndpointSecurityClientTest
from Source/santad/EventProviders/ to Source/common/es/ so they
can be shared with santanetd.

Remove unused Metrics.h include from SNTEndpointSecurityClientTest.
