+++ Source/common/es/SNTEndpointSecurityClient.h
@@ -12,7 +12,7 @@
 ///    See the License for the specific language governing permissions and
 ///    limitations under the License.
 
-#include "Source/santad/EventProviders/SNTEndpointSecurityClientBase.h"
+#include "Source/common/es/SNTEndpointSecurityClientBase.h"
 
 /// This should be treated as an Abstract Base Class and not directly instantiated
 @interface SNTEndpointSecurityClient : NSObject <SNTEndpointSecurityClientBase>
