+++ Source/common/es/SNTEndpointSecurityClientBase.h
