+++ Source/common/es/SNTEndpointSecurityEventHandler.h
