+++ Source/santad/EventProviders/FAAPolicyProcessor.h
@@ -33,9 +33,9 @@
 #include "Source/common/SantaVnode.h"
 #include "Source/common/es/Enricher.h"
 #include "Source/common/es/Message.h"
+#import "Source/common/es/SNTEndpointSecurityEventHandler.h"
 #include "Source/common/faa/WatchItemPolicy.h"
 #include "Source/santad/EventProviders/RateLimiter.h"
-#import "Source/santad/EventProviders/SNTEndpointSecurityEventHandler.h"
 #include "Source/santad/Logs/EndpointSecurity/Logger.h"
 #include "Source/santad/Metrics.h"
 #import "Source/santad/SNTDecisionCache.h"
