+++ Source/santad/EventProviders/SNTEndpointSecurityProcessFileAccessAuthorizer.h
@@ -18,10 +18,10 @@
 
 #include "Source/common/es/ESMetricsObserver.h"
 #include "Source/common/es/EndpointSecurityAPI.h"
+#import "Source/common/es/SNTEndpointSecurityClient.h"
+#import "Source/common/es/SNTEndpointSecurityEventHandler.h"
 #include "Source/common/faa/WatchItems.h"
 #include "Source/santad/EventProviders/FAAPolicyProcessor.h"
-#import "Source/santad/EventProviders/SNTEndpointSecurityClient.h"
-#import "Source/santad/EventProviders/SNTEndpointSecurityEventHandler.h"
 
 @interface SNTEndpointSecurityProcessFileAccessAuthorizer
     : SNTEndpointSecurityClient <SNTEndpointSecurityEventHandler,
