+++ Source/santad/EventProviders/SNTEndpointSecurityProcessFileAccessAuthorizer.mm
@@ -24,8 +24,8 @@
 #import "Source/common/SNTLogging.h"
 #include "Source/common/SantaCache.h"
 #include "Source/common/SantaSetCache.h"
+#include "Source/common/es/SNTEndpointSecurityEventHandler.h"
 #include "Source/common/faa/WatchItemPolicy.h"
-#include "Source/santad/EventProviders/SNTEndpointSecurityEventHandler.h"
 
 using santa::FAAPolicyProcessor;
 using santa::IterateProcessPoliciesBlock;
