+++ Source/santad/EventProviders/SNTEndpointSecurityRecorder.h
@@ -17,8 +17,8 @@
 #include "Source/common/es/ESMetricsObserver.h"
 #include "Source/common/es/EndpointSecurityAPI.h"
 #include "Source/common/es/Enricher.h"
+#import "Source/common/es/SNTEndpointSecurityEventHandler.h"
 #import "Source/santad/EventProviders/AuthResultCache.h"
-#import "Source/santad/EventProviders/SNTEndpointSecurityEventHandler.h"
 #import "Source/santad/EventProviders/SNTEndpointSecurityTreeAwareClient.h"
 #include "Source/santad/Logs/EndpointSecurity/Logger.h"
 #include "Source/santad/ProcessTree/process_tree.h"
