+++ Source/santad/EventProviders/SNTEndpointSecurityTreeAwareClient.h
@@ -12,7 +12,7 @@
 /// See the License for the specific language governing permissions and
 /// limitations under the License.
 
-#include "Source/santad/EventProviders/SNTEndpointSecurityClient.h"
+#include "Source/common/es/SNTEndpointSecurityClient.h"
 
 #include "Source/common/es/ESMetricsObserver.h"
 #include "Source/common/es/EndpointSecurityAPI.h"
