+++ Source/santad/SantadTest.mm
@@ -32,8 +32,8 @@
 #include "Source/common/TestUtils.h"
 #include "Source/common/es/Message.h"
 #include "Source/common/es/MockEndpointSecurityAPI.h"
+#import "Source/common/es/SNTEndpointSecurityClient.h"
 #import "Source/santad/EventProviders/SNTEndpointSecurityAuthorizer.h"
-#import "Source/santad/EventProviders/SNTEndpointSecurityClient.h"
 #import "Source/santad/Metrics.h"
 #include "Source/santad/ProcessControl.h"
 #import "Source/santad/SNTDatabaseController.h"
